Moved AWS unit specs to spec/unit/aws/
