+++ spec/e2e/task.spec.js
@@ -1,0 +1,4 @@
+var setup = require('./setup'),
+Task = require(setup.path('aws', 'taskServiceManager.js'));
+
+module.exports = Task(setup.getEcs());
