+++ spec/unit/aws/ec2-mock.js
