+++ spec/unit/util.spec.js
@@ -11,14 +11,4 @@ describe('utility functions', function () {
   it('should have a function that quotes a supplied argument', function () {
     expect(util.quote('booya')).equal('"booya"');
   });
-
-  it('should have a function that makes an AWS key/value filter',
-  function () {
-    expect(util.makeAWSFilter('t', 1)).to.be;
-  });
-
-  it('should have a function that makes an AWS VPC filter',
-  function () {
-    expect(util.makeAWSFilter(1)).to.be;
-  });
 });
